requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Computer Vision & ML
ultralytics==8.4.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# Data Validation
pydantic==2.5.3

# PyTorch CPU-only (much smaller than CUDA version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
